requirements.txt
instagrapi
moviepy
anthropic
python-dotenv
Pillow
fastapi
uvicorn[standard]
